5273-app - BPartner REST-API
* WIP implementing BL to read and write (json)bpartner composites, incl pagination
* move BPartnerComposite stuff (= sortof bpartner-domain) to dedicated package in de.metas.business
* distinguish between BPartnerComposite ("domain model") and BPartnerInfo (glorified ID-triplet)
https://github.com/metasfresh/metasfresh/issues/5273

+++ de.metas.business.rest-api-impl/src/main/java/de/metas/rest_api/bpartner/impl/MockDataUtil.java
@@ -28,7 +28,7 @@ import lombok.experimental.UtilityClass;
 @UtilityClass
 public class MockDataUtil
 {
-	private static long metasfreshIdCounter = 1;
+	private static int metasfreshIdCounter = 1;
 
 	public MetasfreshId nextMetasFreshId()
 	{
